requirements.txt
Flask==2.2.5
Flask-Session==0.4.0
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==2.0.0
python-dotenv==1.0.1
cryptography==42.0.5
openai==0.28.1
python-docx==1.1.0
pdfminer.six==20250506
gunicorn==23.0.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
supabase>=2.6.0
mammoth==1.10.0
PyMuPDF==1.26.3
pdfrw==0.4
google-cloud-texttospeech==2.27.0
langdetect==1.0.9
pillow==11.3.0
sendgrid==6.10.0
python-docx>=1.0.0
bleach==6.1.0

mutagen>=1.47.0
